nest attributes in resources
