+++ python/lib/sift_client/resources/__init__.py
@@ -163,9 +163,9 @@ from sift_client.resources.ingestion import (
 )
 from sift_client.resources.jobs import JobsAPIAsync
 from sift_client.resources.ping import PingAPIAsync
-from sift_client.resources.principal_attributes import PrincipalAttributesAPIAsync
+from sift_client.resources.access_control.principal_attributes import PrincipalAttributesAPIAsync
+from sift_client.resources.access_control.resource_attributes import ResourceAttributesAPIAsync
 from sift_client.resources.reports import ReportsAPIAsync
-from sift_client.resources.resource_attributes import ResourceAttributesAPIAsync
 from sift_client.resources.rules import RulesAPIAsync
 from sift_client.resources.runs import RunsAPIAsync
 from sift_client.resources.tags import TagsAPIAsync
