+++ python/lib/sift_client/resources/access_control/principal_attributes.py
