+++ python/lib/sift_client/resources/access_control/resource_attributes.py
